+++ src/main/java/hhplus/ecommerce/order/domain/service/OrderService.java
@@ -3,13 +3,13 @@ package hhplus.ecommerce.order.domain.service;
 import hhplus.ecommerce.balance.domain.entity.Balance;
 import hhplus.ecommerce.balance.domain.repository.BalanceRepository;
 import hhplus.ecommerce.exception.MemberNotFoundException;
-import hhplus.ecommerce.order.controller.dto.OrderAppRequest;
+import hhplus.ecommerce.order.domain.dto.OrderAppRequest;
 import hhplus.ecommerce.order.domain.entity.Order;
 import hhplus.ecommerce.order.domain.entity.OrderItem;
-import hhplus.ecommerce.order.domain.entity.OrderSheet;
-import hhplus.ecommerce.order.domain.entity.OrderSheetItem;
+import hhplus.ecommerce.ordersheet.domain.entity.OrderSheet;
+import hhplus.ecommerce.ordersheet.domain.entity.OrderSheetItem;
 import hhplus.ecommerce.order.domain.repository.OrderRepository;
-import hhplus.ecommerce.order.domain.repository.OrderSheetRepository;
+import hhplus.ecommerce.ordersheet.domain.repository.OrderSheetRepository;
 import hhplus.ecommerce.order.mapper.OrderMapper;
 import hhplus.ecommerce.product.domain.entity.Product;
 import hhplus.ecommerce.product.domain.entity.ProductOption;
