+++ src/main/java/hhplus/ecommerce/ordersheet/controller/dto/OrderSheetRequestDto.java
@@ -1,4 +1,4 @@
-package hhplus.ecommerce.order.controller.dto;
+package hhplus.ecommerce.ordersheet.controller.dto;
 
 import java.util.List;
