+++ src/main/java/hhplus/ecommerce/ordersheet/domain/entity/OrderSheet.java
@@ -1,4 +1,4 @@
-package hhplus.ecommerce.order.domain.entity;
+package hhplus.ecommerce.ordersheet.domain.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
